requirements.txt
streamlit==0.67.1
spacy==2.3.2
st-annotated-text